Use of RObject as a parameter type
Though it would have been possible to create an RObject
JobParam with the changes to scripts.py, there was no
helper class to make it nice and tidy. This change adds
an omero.scripts.Object type and some base examples.

What's missing from this commit is run-time checking
that the actual passed value matches the default. In fact,
we may want to use the "values" collection, so that we can
say "takes OriginalFile or FileAnnotation types". This will
need some discussion.