Caching the DetailsWrapper. (See #8858)
We are also removing the "special" getDetails() implementation that is
currently present on the ExperimenterWrapper due to potential collisions
with the new implementation.  Furthermore, it makes direct modifications
to the underlying omero.model.Details object which is forbidden in all
but the most specific use cases where no other alternative is available.