Enforce top-level user-directory in managed repo
The decision was to maintain a top-level user directory
which has special permission handling. The directory
(of the form `%name%_%id%`) will be placed in the "user"
group so that it is visible from all queries.

NB: This does not yet fix the rest of the code that assumes
writing to the top directory is permissible!