Fix failing test_import.py naming integration tests
For some reason, depending on the method used to run the tests, the Object:id
line is not always at the same position in the stderr. This commit applies the
same strategy as the one used in test.integration.library.import_image() i.e.
loop over the stderr lines (in reverse order) and break when a match is found.