Add tests for password checks after user creation/password modification
Testing a connection with the valid password is not sufficient as it does not
account for the case of empty passwords (where any input password is
accepted). This commit adds an extra check for a failing connection with a
random password. The explicit case where --no-password is passed to user add
is handled in a separate test.