binary_access: Adjust config explanation and implementation
There was some confusion about the ordering of the binary_access
property which has hopefully now be clarified in etc/omero.properties.

Also, it was unclear what the behavior between group & server
settings. This has been simplified by making the implementation
simply default to "pessimistic" behavior. This means that for
plates to be downloadable at all, there *must* be a server-level
configuration.